Revert "refactor: 피드 조회시에 스터디 이름도 반환"
This reverts commit 34b4907fbe6c768e0765d9bf849771f9d349cada.
